+++ include/invader/tag/parser/compile/model.hpp
@@ -9,13 +9,13 @@ namespace Invader::Parser {
     struct GBXModel;
     struct GBXModelGeometryPart;
     bool uncache_model_markers(GBXModel &model, bool fix);
-    bool regenerate_missing_model_vertices(GBXModelGeometryPart &part, bool fix);
+    bool regenerate_missing_model_vertices(GBXModelGeometryPart &part, GBXModel &model, bool fix);
     bool regenerate_missing_model_vertices(GBXModel &model, bool fix);
     
     struct Model;
     struct ModelGeometryPart;
     bool uncache_model_markers(Model &model, bool fix);
-    bool regenerate_missing_model_vertices(ModelGeometryPart &part, bool fix);
+    bool regenerate_missing_model_vertices(ModelGeometryPart &part, Model &model, bool fix);
     bool regenerate_missing_model_vertices(Model &model, bool fix);
     
     enum MaxCompressedModelNodeIndex : std::uint8_t {